support: Compile mod info into support package and use a temp file when creating the package (#545)
My theory still remains that users are uploading the zip before it's
done being constructed (or it fails for some reason).

So I switched to making it as a temp file, then copying that temp file
over if everything succeeds. This should hopefully help...

+++ src-tauri/src/commands/features/mods.rs
@@ -780,7 +780,9 @@ pub async fn uninstall_mod(
     .join("mods")
     .join(&source_name)
     .join(&mod_name);
-  std::fs::remove_dir_all(mod_dir)?;
+  if mod_dir.exists() {
+    std::fs::remove_dir_all(mod_dir)?;
+  }
   config_lock
     .uninstall_mod(game_name, mod_name, source_name)
     .map_err(|_| CommandError::GameFeatures("Unable to uninstall mod".to_owned()))?;
